Rewrite gitconfig.example to be my base for gitconfig
While the .example file was fantastic in holman's public repo, it
doesn't make a lot of sense in my own repo. Instead, since gitconfig is
an unversioned file, let's use it as a template for future clones of
this repo. All of my aliases and my preferred git settings are present.
Furthermore, `rake install` will now copy this over to a gitconfig
assuming that gitconfig doesn't exist already.