Only look for install.sh scripts in dotfile root dirs.
This follows the same pattern as script/bootstrap by explicitly only
looking under dotfile root dirs for the install.sh scripts.

Looking in the parent directory can find unrelated install.sh not intended
to be executed (which is what happened to me :), when I was in .dotfiles
and ran `scripts/install`.