Reload tmux conf after update
After updating or when calling the !reload command, this file is loaded from
the zshrc and will cause to running tmux instances to reload the configuration.
This might lead to zombie states, though, as the cumulative effect of loading
the old conf first and then loading the new conf is the effect of this script.