Do not run bootstrapping automatically after update
When opening a new terminal session, it is ok to run the upgrade
process. Yet, it is problematic to run the bootstrapping as it requires
user input to choose if files should be overridden or backuped. When
the user opened the terminal unaware of the automatic update, he might
already have typed input.