todo 2.0
I've been using your approach to Inbox Zero lately, and I've discovered a little twist. If there's an email that needs a reply or has instructions for something that I need done, I can just drag it to the desktop. It creates a copy, so the original message is still in my mail folder. I can delete the copy from my desktop as soon as I'm done with it. The same goes for URLs. If I need to work on an issue or review a pull request, I simply drag the link to my desktop / todo list, and when I'm ready to work on that item, it's just a double-click away.