Handle completely base16 with vim, iterm and Tmux
After hours of hard work, I managed to make this trio work seamlessly
together. However, this works perfectly in MacVim and not in Terminal. In
fact, and it seems "normal" by the maintainer of base16, the only way to
change the vim base16 colorscheme is through the shell utility.
In terminal Vim, we cannot have a different base16 colorscheme other than
the one set in iTerm.

This commit includes also a `color` utility with autocompletion to easily
switch between the base16 colorschemes + some niceties that allow MacVim to
grab automatically the new shell colorscheme when it reaches the focus (
through the `FocusGained` auto command)