Huge refactoring of rcfiles + deferring mechanism
This commit exploits the built-in features of vim by moving bits of code
in their appropriate vim folders :
- Filetype detection code : in ftdetect/ folder
- Related filetype local settings : in ftplugin/ related folder
- Autoload code : Code to be sourced only when called

This commit also adds deferring mechanism in the autoload folder, in
addition to the autoload folder.
Basically, in our rcfiles, we can call the defer#defer method and pass
it a function, usually another autoloaded function, so that it is called
after startup.