Sets up light and dark mode for Vim/Kitty.
Generally, this works by asking macOS if dark mode is on at start up of
Kitty (via zsh starting) and Vim. This means they won't interaact like
other mcaOS apps that change immediately when daark mode is turned on
(or off), but, if needed the capability to set light and dark mode
mnaually was added.

This uses the light and dark themes from the [tomorrow
theme](https://github.com/chriskempson/tomorrow-theme); specifically,
Tomorrow (for light) and Tomorrow Night Eighties for dark mode.

* Adds `zshenv` file with alias for checking macOS if using dark mode
  (`darkMode`)
  * This needs to be in `zshenv` to be available in non-interactive
    shells (which is how Vim commands work)
* Uses `darkMode` check in `zshrc` config to use set Kitty theme
* Creates scripts for setting Kitty to light and dark mode
* Adds functions to nvim for light and dark mode, makes them callable
  via a leader key combo, and checks the `darkMode` alias at startup to
  set colorscheme based on system-level dark mode setting.