fix(nvim): remove special lsp callback handlers (#20)
With a recent nvim (lsp) update the number of arguments to the callback
handler changed. Since this code's difference to the original code is
to modify the behavior in the presence of non-saved changes, we opt to
use the original handlers instead of overriding them. This saves us the
hassle of maintaining custom handlers.